Pull ScadrLoaderClient out.
